Disable CodeNarc
Not terribly familiar with CodeNarc, and it is throwing an exception when
trying to build lazybones-app. To facilitate working on the bug at hand,
it is being disabled until it can properly be addressed.

Build completes, but there are a number of integration test failures.
Of course, that lends to the question as to why the build completes as
successful when there are test failures?