Upgrade wrapper version: 2.9 -> 3.5.1
IntelliJ is unable to work with version 2.9 of Gradle and requested
the use of at least version 3.0; since there is an open ticket to
upgrade to the latest Gradle version, this commit only attempts to
upgrade to the point where IntelliJ is usable. Confirmed that the
build still works as before (including integration test failures)
while updating to the latest v3.x release of Gradle.